Adding locked
When some time has passed since the release of a program, new dependencies can prevent cargo from building that application. The `--locked` flag forces cargo to build the application with the original dependencies in which it was released.
Changing to a block of code allows you to copy it with the button on the right.